Use `%zu` and `%td` in printf format strings for size_t and ptrdiff_t
Had to include stddef.h to make ptrdiff_t available in OpenBSD 6.0,
compilation failed with "error: 'ptrdiff_t' undeclared".

65c1a9e8-3565-11eb-b330-4f45262dc9b5

+++ src/uuid.c
@@ -396,9 +396,9 @@ char *uuid_date(char *dest, const char *uuid)
 	strftime(dest, DATE_LENGTH, "%Y-%m-%dT%H:%M:%S", tm);
 	p = dest + strlen(dest);
 	if (p - dest != 19) {
-		fprintf(stderr, "%s: %lu: Invalid date length," /* gncov */
+		fprintf(stderr, "%s: %td: Invalid date length," /* gncov */
 		                " should be 19\n", progname,
-		                (unsigned long)(p - dest)); /* gncov */
+		                (ptrdiff_t)(p - dest)); /* gncov */
 		return NULL; /* gncov */
 	}
 	snprintf(p, DATE_LENGTH - 19 + 1, ".%07uZ", nano);
